pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brlcad-mcp"
version = "0.1.0"
description = "Natural language interface for BRL-CAD solid modeling via MCP."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "BRL-CAD MCP Contributors" }]

dependencies = [
    "mcp",
    "pydantic",
    "python-dotenv",
    "langchain-openai",
    "langchain-mcp-adapters",
    "langgraph",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[project.scripts]
brlcad-mcp = "brlcad_mcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/brlcad_mcp"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
